+++ src/main/java/pl/asie/computronics/reference/Names.java
@@ -10,6 +10,8 @@ public class Names {
 	public static final String FZ_ChargeConductor = "charge_conductor";
 	public static final String MFR_DSU = "deep_storage_unit";
 	public static final String Railcraft_ElectricGrid = "electric_tile";
+	public static final String Railcraft_LimiterTrack = "limiter_track";
+	public static final String Railcraft_LocomotiveTrack = "locomotive_track";
 	public static final String Railcraft_ReceiverBox = "receiver_box";
 	public static final String Railcraft_RoutingDetector = "routing_detector";
 	public static final String Railcraft_RoutingSwitch = "routing_switch";