+++ changelog.md
@@ -9,7 +9,7 @@
   * ADDED: More Waila Integration
             - Tape Drives now show their current state in Waila
   * CHANGED: Moved to OpenComputers 1.4
-  * CHANGED: MAJOR code refactor, file size should be a LOT smaller.
+  * CHANGED: MAJOR refactor of integration code.
   * FIXED: A pretty serious bug with the Cipher Block
   * FIXED: Waila integration
   * FIXED: Tape Drive's getState not properly returning the current state.